requirements.txt
# HTML Parsing
beautifulsoup4>=4.9.3
lxml>=4.6.0

# Async HTTP and file operations
aiohttp>=3.8.0
aiofiles>=0.8.0

# Browser automation for page fetching
undetected-chromedriver>=3.5.0
selenium>=4.0.0

# Environment variable management (optional but recommended)
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
